+++ apps/web/src/components/Shared/Embed/Space/SpacePlayer.tsx
@@ -1,5 +1,19 @@
+import { getLensAccessToken, getLensMessage, getMessage } from '@huddle01/auth';
+import {
+  useAudio,
+  useHuddle01,
+  useLobby,
+  useMeetingMachine,
+  usePeers,
+  useRoom
+} from '@huddle01/react/hooks';
 import type { Publication } from '@lenster/lens';
+import { useState } from 'react';
 import type { FC } from 'react';
+import { useEffectOnce, useUpdateEffect } from 'usehooks-ts';
+import { useAccount, useSignMessage } from 'wagmi';
+import { useDisplayName } from '@huddle01/react/app-utils';
+import { useAppStore } from 'src/store/app';
 
 interface SpacePlayerProps {
   publication: Publication;
@@ -7,9 +21,91 @@ interface SpacePlayerProps {
 }
 
 const SpacePlayer: FC<SpacePlayerProps> = ({ publication, space }) => {
+  const currentProfile = useAppStore((state) => state.currentProfile);
+  const [accessToken, setAccessToken] = useState('');
+  const { initialize, isInitialized } = useHuddle01();
+  const { joinLobby, isLobbyJoined } = useLobby();
+  const {
+    fetchAudioStream,
+    stopProducingAudio,
+    produceAudio,
+    stream: micStream
+  } = useAudio();
+  const { joinRoom, leaveRoom } = useRoom();
+  const { setDisplayName } = useDisplayName();
+  const { peerIds: peers } = usePeers();
+  const { state } = useMeetingMachine();
+  const { address } = useAccount();
   const { metadata } = publication;
 
-  return <div className="p-5">gm</div>;
+  const { signMessage } = useSignMessage({
+    onSuccess: async (data) => {
+      const token = await getLensAccessToken(data, address as string);
+      console.log('data', token);
+      setAccessToken(token.accessToken);
+    }
+  });
+
+  useEffectOnce(() => {
+    initialize('KL1r3E1yHfcrRbXsT4mcE-3mK60Yc3YR');
+    const getAccessToken = async () => {
+      const msg = await getLensMessage(address as string);
+      signMessage({ message: msg.message });
+    };
+
+    getAccessToken();
+  });
+
+  useUpdateEffect(() => {
+    if (isInitialized && accessToken) {
+      joinLobby('xuv-ajbu-hrw', accessToken);
+    }
+  }, [isInitialized, accessToken]);
+
+  useUpdateEffect(() => {
+    if (fetchAudioStream.isCallable) {
+      setDisplayName(currentProfile?.id);
+      fetchAudioStream();
+    }
+  }, [isLobbyJoined]);
+
+  return (
+    <div className="p-5">
+      {accessToken ? (
+        <>
+          <h2 className="text-2xl">Room State</h2>
+          <h3 className="break-words">{JSON.stringify(state.value)}</h3>
+          <h2 className="text-2xl">DisplayName</h2>
+          <div className="break-words">
+            {JSON.stringify(state.context.displayName)}
+          </div>
+          <h2 className="text-2xl">Peers</h2>
+          <div className="break-words">{JSON.stringify(peers)}</div>
+          <button disabled={!joinRoom.isCallable} onClick={joinRoom}>
+            Join
+          </button>
+          <br />
+          <button disabled={!leaveRoom.isCallable} onClick={leaveRoom}>
+            Leave
+          </button>
+          <br />
+          <button
+            disabled={!produceAudio.isCallable}
+            onClick={() => produceAudio(micStream)}
+          >
+            Talk
+          </button>
+          <br />
+          <button
+            disabled={!stopProducingAudio.isCallable}
+            onClick={stopProducingAudio}
+          >
+            Mute
+          </button>
+        </>
+      ) : null}
+    </div>
+  );
 };
 
 export default SpacePlayer;
