--- apps/web/src/pages/spaces.tsx
@@ -1,3 +1,0 @@
-import Spaces from '@components/Spaces';
-
-export default Spaces;
